FDG-10651 Adjust interval
